Formalize errors using gRPC status and codes API
This integrates the status and codes packages in the gRPC which are
independent of gRPC itself as a transport and protocol. The reasoning
is to use well-defined codes and existing protobuf definition which
enables seamless interop of propagating codes to gRPC interfaces.

Note, this is a backwards compatible change.

Fix #4

Signed-off-by: Byron Ruth <b@devel.io>